Refactor ipynb stuff into a new module, minor lint fixes
