+++ tests/test_ipynb.py
@@ -3,7 +3,7 @@ from collections import namedtuple
 import pytest
 
 from pytest_split.algorithms import Algorithms
-from pytest_split.plugin import _reorganize_broken_up_ipynbs
+from pytest_split.ipynb_compatibility import _reorganize_broken_up_ipynbs
 
 item = namedtuple("item", "nodeid")
 
@@ -30,22 +30,22 @@ class TestIPyNb:
         }
         items = [item(x) for x in durations.keys()]
         algo = Algorithms[algo_name].value
-        g1, g2, g3 = algo(splits=3, items=items, durations=durations)
+        groups = algo(splits=3, items=items, durations=durations)
 
-        assert g1.selected == [
+        assert groups[0].selected == [
             item(nodeid="temp/nbs/test_1.ipynb::Cell 0"),
             item(nodeid="temp/nbs/test_1.ipynb::Cell 1"),
             item(nodeid="temp/nbs/test_1.ipynb::Cell 2"),
             item(nodeid="temp/nbs/test_2.ipynb::Cell 0"),
             item(nodeid="temp/nbs/test_2.ipynb::Cell 1"),
         ]
-        assert g2.selected == [
+        assert groups[1].selected == [
             item(nodeid="temp/nbs/test_2.ipynb::Cell 2"),
             item(nodeid="temp/nbs/test_2.ipynb::Cell 3"),
             item(nodeid="temp/nbs/test_3.ipynb::Cell 0"),
             item(nodeid="temp/nbs/test_3.ipynb::Cell 1"),
         ]
-        assert g3.selected == [
+        assert groups[2].selected == [
             item(nodeid="temp/nbs/test_3.ipynb::Cell 2"),
             item(nodeid="temp/nbs/test_3.ipynb::Cell 3"),
             item(nodeid="temp/nbs/test_3.ipynb::Cell 4"),
@@ -54,8 +54,8 @@ class TestIPyNb:
             item(nodeid="temp/nbs/test_4.ipynb::Cell 2"),
         ]
 
-        _reorganize_broken_up_ipynbs(g1, items)
-        assert g1.selected == [
+        _reorganize_broken_up_ipynbs(groups[0], items)
+        assert groups[0].selected == [
             item(nodeid="temp/nbs/test_1.ipynb::Cell 0"),
             item(nodeid="temp/nbs/test_1.ipynb::Cell 1"),
             item(nodeid="temp/nbs/test_1.ipynb::Cell 2"),
@@ -65,8 +65,8 @@ class TestIPyNb:
             item(nodeid="temp/nbs/test_2.ipynb::Cell 3"),
         ]
 
-        _reorganize_broken_up_ipynbs(g2, items)
-        assert g2.selected == [
+        _reorganize_broken_up_ipynbs(groups[1], items)
+        assert groups[1].selected == [
             item(nodeid="temp/nbs/test_3.ipynb::Cell 0"),
             item(nodeid="temp/nbs/test_3.ipynb::Cell 1"),
             item(nodeid="temp/nbs/test_3.ipynb::Cell 2"),
@@ -74,20 +74,9 @@ class TestIPyNb:
             item(nodeid="temp/nbs/test_3.ipynb::Cell 4"),
         ]
 
-        _reorganize_broken_up_ipynbs(g3, items)
-        assert g3.selected == [
+        _reorganize_broken_up_ipynbs(groups[2], items)
+        assert groups[2].selected == [
             item(nodeid="temp/nbs/test_4.ipynb::Cell 0"),
             item(nodeid="temp/nbs/test_4.ipynb::Cell 1"),
             item(nodeid="temp/nbs/test_4.ipynb::Cell 2"),
         ]
-
-
-if __name__ == "__main__":
-    import warnings
-
-    warnings.filterwarnings("ignore")
-
-    import pytest
-
-    args = "test_ipynb.py"
-    retcode = pytest.main(args.split())
